micro refacto of /common  - added tool module to declare export tool.component  - moved safe content from /common/pipe/safe-content to /common/safe-content  - added /asset module/service  - ds commons module just does import export of commons sub modules  - clean up indexes
