+++ src/app/commons/asset/asset.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+// import { environment } from '../../../environments/environment';
+
+@NgModule({
+	imports: [],
+	providers: [],
+})
+export class DsAssetModule {}
