+++ src/app/commons/asset/asset.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class DsAssetService {
+
+}
