+++ src/app/commons/asset/index.ts
@@ -1,0 +1,2 @@
+export * from './asset.module';
+export * from './asset.service';
