+++ src/app/commons/index.ts
@@ -1,1 +1,8 @@
 export * from './commons.module';
+export * from './asset';
+export * from './code-tabs';
+export * from './doc-api';
+export * from './page';
+export * from './safe-content';
+export * from './structure';
+export * from './tool';
