+++ src/app/commons/safe-content/index.ts
