+++ src/app/commons/safe-content/safe-content.pipe.ts
