+++ src/app/commons/structure/examples-browser/examples-browser.module.ts
@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { LuSelectModule, LuOptionModule, LuInputDisplayerModule } from '@lucca-front/ng';
 import { ExamplesBrowserComponent } from './examples-browser.component';
 import { DsCodeTabsModule } from '../../code-tabs/code-tabs.module';
-import { DsSafeContentModule } from '../../pipes/safe-content';
+import { DsSafeContentModule } from '../../safe-content';
 import { DsExampleService } from './example.service';
 
 @NgModule({
