+++ src/app/commons/tool/index.ts
@@ -1,0 +1,2 @@
+export * from './tool.component';
+export * from './tool.module';
