+++ src/app/commons/tool/tool.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+
+import { ToolComponent } from './tool.component';
+
+@NgModule({
+	imports: [
+	],
+	declarations: [
+		ToolComponent
+	],
+	exports: [
+		ToolComponent,
+	],
+	entryComponents: [
+		ToolComponent,
+	]
+})
+export class DsToolModule { }
